requirement.txt
streamlit==1.40.2
google-ai-generativelanguage==0.6.10
google-generativeai==0.8.3
python-dotenv==1.0.1